backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# LLM & AI
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16
langgraph==0.0.20
google-generativeai==0.3.2

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.2.0

# NLP & Translation
spacy==3.7.2
langdetect==1.0.9
deep-translator==1.11.4

# File Processing
python-multipart==0.0.6
aiofiles==23.2.1
openpyxl==3.1.2

# Email & Communication
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.26.0
websockets==12.0
slowapi==0.1.9

# Caching & Task Queue
redis==5.0.1
celery==5.3.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Apex SaaS Framework
apex-saas-framework

# Async Database
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.0
